VisualFL/requirements.txt
attr>=0.3.1
grpcio>=1.33.2,!=1.34.0
aiohttp>=3.7,<3.8
loguru>=0.5
protobuf==3.14.0
jsonschema==3.2.0
PyYAML>=5.3.1
click==7.1.2
paddlepaddle==1.8.5

# enhance
visualdl==2.0.5